requirements.txt
beautifulsoup4==4.12.3
blinker==1.6.2   
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
neo4j==5.18.0
python-decouple==3.8
python-steam-api==1.2.2
pytz==2024.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==2.3.4